refactor(ui): Rename TableFilters -> FiltersDropdown
