--- packages/ui/src/ui/table-filters/index.js
@@ -1,4 +1,0 @@
-import { tableFiltersContainer } from './table-filters.container';
-import { TableFilters as TableFiltersComponent } from './table-filters';
-
-export const TableFilters = tableFiltersContainer(TableFiltersComponent);
